==> 1.Data_Generation.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime, timedelta

ROOM_TYPES = ["Standard", "Deluxe", "Executive", "Suite"]

ROOM_PRICES = {
    "Standard": 2000,
    "Deluxe": 4000,
    "Executive": 8000,
    "Suite": 12000
}

def generate_random_date(start_date, end_date):
    """
    Generates a random date between start_date and end_date
    """
    delta_days = (end_date - start_date).days
    random_days = np.random.randint(0, delta_days)
    return start_date + timedelta(days=random_days)


START_DATE = datetime(2024, 1, 1)
END_DATE = datetime(2025, 12, 31)

NUM_RECORDS = 10000

records = []

for booking_id in range(1, NUM_RECORDS + 1):

    room_type = np.random.choice(ROOM_TYPES)

    check_in = generate_random_date(START_DATE, END_DATE)
    stay_duration = np.random.randint(1, 8)  # 1 to 7 days
    check_out = check_in + timedelta(days=stay_duration)

    room_service_cost = np.random.randint(0, 3000)

    total_bill = (
        ROOM_PRICES[room_type] * stay_duration
        + room_service_cost
    )

    records.append([
        booking_id,
        room_type,
        check_in,
        check_out,
        stay_duration,
        ROOM_PRICES[room_type],
        room_service_cost,
        total_bill
    ])


columns = [
    "Booking_id",
    "Room_type",
    "Check_in_date",
    "Check_out_date",
    "Stay_duration",
    "Room_price",
    "Room_service_cost",
    "Total_bill"
]

df = pd.DataFrame(records, columns=columns)


df.head()

df.info()

df.to_csv("hotel_bookings.csv", index=False)
