==> 2.Data Manipulation Using Pandas.ipynb <==
import pandas as pd

df = pd.read_csv("hotel_bookings.csv")

df.head()

df.shape

df.columns

df.dtypes

df["Check_in_date"] = pd.to_datetime(df["Check_in_date"])
df["Check_out_date"] = pd.to_datetime(df["Check_out_date"])

df.head(10)

df.tail()

Filtering data

# Only suites
df[df["Room_type"] == "Suite"]

# High-value bookings
df[df["Total_bill"] > 20000]

# Multiple conditions
df[
    (df["Room_type"] == "Deluxe") &
    (df["Stay_duration"] > 3)
]

# Top customers/revenue Data
df.sort_values(by="Total_bill", ascending=False)

# Statistical analysis
df["Total_bill"].mean() #Calculates the average revenue generated per booking across all hotel stays.

df["Total_bill"].max() #Finds the highest single-booking revenue earned by the hotel.

df["Total_bill"].min() #Identifies the lowest revenue generated from a single booking.

# Revenue by room type
df.groupby("Room_type")["Total_bill"].sum()

# Average stay duration
df.groupby("Room_type")["Stay_duration"].mean()

# Feature Engineering
df["Price_per_day"] = df["Total_bill"] / df["Stay_duration"]

df.to_csv("Hotel_bookings_updated.csv", index=False)

df.info()
