==> 3.Data_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv("Hotel_bookings_updated.csv")
df

df.head()

df.info()

df.isnull().sum() #Missing values

df["Check_in_date"] = pd.to_datetime(df["Check_in_date"])
df["Check_out_date"] = pd.to_datetime(df["Check_out_date"])

df.info()

df = df[df["Stay_duration"] > 0] #stay duration is possible

df = df[df["Total_bill"] > 0] #Revenue cannot be zero or negative.

df.duplicated().sum() #finding duplicates

df = df[df["Check_out_date"] > df["Check_in_date"]] #logical data validation

#Recalculate Stay Duration
df["Stay_duration"] = (
    df["Check_out_date"] - df["Check_in_date"]
).dt.days

df.describe() #descriptive statistics analysis

df.describe(include='all') #descriptive statistics analysis for both numerical and categorical data

df.to_csv("Hotel_bookings_cleaned.csv", index=False)
